succeed in first running; add 1. gpu 2. save and load shortly 3. testing part
